refactor(account): 拆分 components.py + 清理冗余
- CredPopup/CredLine/AccountRow 移入 components.py（285 行）
- page.py 仅保留 AccountDialog（198 行），与 editor/history 结构一致
- CredLine 改用 account.masked_sessdata/masked_bili_jct，移除 _mask 导入
- account_form.py 移除未使用的 QMessageBox 导入
